+++ lib/arkit_plugin.dart
@@ -34,6 +34,7 @@ export 'package:arkit_plugin/light/arkit_light_type.dart';
 export 'package:arkit_plugin/physics/arkit_physics_body.dart';
 export 'package:arkit_plugin/physics/arkit_physics_body_type.dart';
 export 'package:arkit_plugin/physics/arkit_physics_shape.dart';
+export 'package:arkit_plugin/widget/ar_environment_texturing.dart';
 export 'package:arkit_plugin/widget/arkit_arplane_detection.dart';
 export 'package:arkit_plugin/widget/arkit_configuration.dart';
 export 'package:arkit_plugin/widget/arkit_scene_view.dart';